Allow errors to be ignored using @
PHP has the @-operator for function calls:
@myFunc();
This will set error_reporting() to 0 for the reporting of this error,
aka: allows us to run code that _may_ error __gracefully__ without
having to worry about exceptions and so on.

Apply this to the sockets, because this currently causes the sockets
to fail when parsing 'GET / HTTP/1.1' as this line contains no ':'
to explode on.