Changed default hasPaid to true
The config by default is set to immediately reserve ISIC cards, which corresponds to hasPaid being true, so this is a morge logical choice. It only really affects ISIC cards which already are in the database (which also got immediately ordered). New Isic Cards always have the hasPaid value explicitly set upon creation.